+++ CHANGELOG.md
@@ -72,6 +72,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Windows binaries no longer require `vcruntime140.dll` to be installed
   (normally through Visual Studio.)
 
+* `jj branch rename` no longer shows a warning in colocated repos.
+
 ## [0.19.0] - 2024-07-03
 
 ### Breaking changes